+++ src/Auth/Auth.php
@@ -6,10 +6,9 @@ use Hyvor\Internal\Component\Component;
 use Hyvor\Internal\Component\InstanceUrlResolver;
 use Hyvor\Internal\InternalApi\Exceptions\InternalApiCallFailedException;
 use Hyvor\Internal\InternalApi\InternalApi;
-use Hyvor\Internal\InternalApi\InternalApiMethod;
-use Illuminate\Http\Request;
 use Illuminate\Support\Collection;
 use Symfony\Component\HttpFoundation\RedirectResponse;
+use Symfony\Component\HttpFoundation\Request;
 
 /**
  * @phpstan-import-type AuthUserArray from AuthUser
@@ -21,7 +20,7 @@ class Auth implements AuthInterface
 
     public function __construct(
         private InternalApi $internalApi,
-        private InstanceUrlResolver $instanceUrlResolver
+        private InstanceUrlResolver $instanceUrlResolver,
     ) {
     }
 
@@ -30,9 +29,12 @@ class Auth implements AuthInterface
      */
     public function check(string $cookie): false|AuthUser
     {
+        if (empty($cookie)) {
+            return false;
+        }
+
         $response = $this->internalApi->call(
             Component::CORE,
-            InternalApiMethod::POST,
             '/auth/check',
             [
                 'cookie' => $cookie
@@ -45,43 +47,38 @@ class Auth implements AuthInterface
         return is_array($user) ? AuthUser::fromArray($user) : false;
     }
 
+    /**
+     * @param string $page page in core to redirect to
+     * @param string|null $redirect absolute URL to redirect back to
+     */
     private function redirectTo(
         string $page,
-        ?string $redirectPage = null
+        null|string $redirect
     ): RedirectResponse {
         $pos = strpos($page, '?');
         $placeholder = $pos === false ? '?' : '&';
 
-        /** @var Request $request */
-        $request = request();
-
-        if ($redirectPage === null) {
-            $redirectPage = $request->getPathInfo();
+        $redirectQuery = '';
+        if (is_string($redirect)) {
+            $redirectQuery = $placeholder . 'redirect=' . urlencode($redirect);
         }
 
-        $redirectUrl = $redirectPage && str_starts_with($redirectPage, 'https://')
-            ? $redirectPage
-            : $request->getSchemeAndHttpHost() . $redirectPage;
-
-        $redirect = $placeholder . 'redirect=' .
-            urlencode($redirectUrl);
-
-        $fullUrl = $this->instanceUrlResolver->publicUrlOfCore() . '/' . $page . $redirect;
+        $fullUrl = $this->instanceUrlResolver->publicUrlOfCore() . '/' . $page . $redirectQuery;
 
         return new RedirectResponse($fullUrl);
     }
 
-    public function login(?string $redirect = null): RedirectResponse
+    public function login(null|string|Request $redirect = null): RedirectResponse
     {
         return $this->redirectTo('login', $redirect);
     }
 
-    public function signup(?string $redirect = null): RedirectResponse
+    public function signup(null|string|Request $redirect = null): RedirectResponse
     {
         return $this->redirectTo('signup', $redirect);
     }
 
-    public function logout(?string $redirect = null): RedirectResponse
+    public function logout(null|string|Request $redirect = null): RedirectResponse
     {
         return $this->redirectTo('logout', $redirect);
     }
@@ -94,9 +91,8 @@ class Auth implements AuthInterface
      */
     protected function getUsersByField(string $field, iterable $values): Collection
     {
-        $response = InternalApi::call(
+        $response = $this->internalApi->call(
             Component::CORE,
-            InternalApiMethod::POST,
             '/auth/users/from/' . $field,
             [
                 $field => (array)$values