+++ CHANGELOG.md
@@ -5,9 +5,12 @@
 - Component::current() is removed, use InternalConfig instead
 - Auth methods are no longer static
 - Auth login, signup, and logout methods now return a Symfony RedirectRepsonse object
+- Auth login, signup, and logout methods now require a string URL to set redirect URL to the current one
 - component label is removed from metrics, add that in Prometheus instead
 - InternalApiMethod is removed, call internal API requests should be POST now
 - PHP 8.4 required
+- AuthFake::enable not binds AuthInterface, not Auth, which means apps needs to type hint AuthInterface in dependency
+  injection
 
 ## 2.1.0 - 2025-03-23